Broad update: cross-vm bridge hardening, pallet upgrades, DEX UI, x3-desktop rebuild, fraud proofs, supply ledger fixes, clinerules additions, x3-lang compiler/VM e2e pipeline, infrastructure dashboards

+++ crates/x3-staking-analytics/src/staking_simulator.rs
@@ -125,12 +125,12 @@ impl StakingSimulator {
             current_balance += monthly_deposits as f64;
 
             let gross_rewards = (current_balance as u128).saturating_sub(
-                initial_balance + (monthly_deposits as u128 * month)
+                initial_balance + (monthly_deposits as u128 * month as u128)
             );
             let commission_cost =
                 (gross_rewards as f64 * (validator_commission / 100.0)) as u128;
             let net_rewards = gross_rewards.saturating_sub(commission_cost);
-            let total_invested = initial_balance + (monthly_deposits as u128 * month);
+            let total_invested = initial_balance + (monthly_deposits as u128 * month as u128);
             let roi = if total_invested > 0 {
                 (net_rewards as f64 / total_invested as f64) * 100.0
             } else {
